+++ modfile.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"io"
 	"os"
@@ -47,7 +48,15 @@ func updateMod(trueTestModules []string, f *modfile.File) error {
 				// add test comment
 				line := fr.Syntax
 				setTest(line, true)
-				lineMods = append(lineMods, lineMod{name: line.Token[0], ver: line.Token[1], coms: line.Comments})
+				if len(line.Token) == 2 {
+					// eg; `github.com/shirou/gopsutil v1.21.5`
+					lineMods = append(lineMods, lineMod{name: line.Token[0], ver: line.Token[1], coms: line.Comments})
+				} else if len(line.Token) == 3 {
+					// eg; `require rsc.io/quote v1.5.2`
+					lineMods = append(lineMods, lineMod{name: line.Token[1], ver: line.Token[2], coms: line.Comments})
+				} else {
+					return errors.New("unrecognised line.Token format")
+				}
 			}
 		}
 	}
@@ -79,6 +88,73 @@ func addTestRequireBlock(f *modfile.File, lineMods []lineMod) {
 			)
 	*/
 
+	findLastRequire := func(f *modfile.File) int {
+		/*
+			The code of this function is mostly file is taken from https://github.com/golang/mod/
+			The license of that repo is found at: https://github.com/golang/mod/blob/v0.3.0/LICENSE
+			and is included below:
+
+				Copyright (c) 2009 The Go Authors. All rights reserved.
+
+				Redistribution and use in source and binary forms, with or without
+				modification, are permitted provided that the following conditions are
+				met:
+
+				   * Redistributions of source code must retain the above copyright
+				notice, this list of conditions and the following disclaimer.
+				   * Redistributions in binary form must reproduce the above
+				copyright notice, this list of conditions and the following disclaimer
+				in the documentation and/or other materials provided with the
+				distribution.
+				   * Neither the name of Google Inc. nor the names of its
+				contributors may be used to endorse or promote products derived from
+				this software without specific prior written permission.
+
+				THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+				"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+				LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+				A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+				OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+				SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+				LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+				DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+				THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+				(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+				OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+		*/
+
+		var pos int
+
+		for i, stmt := range f.Syntax.Stmt {
+			switch stmt := stmt.(type) {
+			case *modfile.Line:
+				if len(stmt.Token) == 0 || stmt.Token[0] != "require" {
+					continue
+				}
+				pos = i
+			case *modfile.LineBlock:
+				if len(stmt.Token) == 0 || stmt.Token[0] != "require" {
+					continue
+				}
+				pos = i
+			}
+		}
+
+		return pos
+	}
+
+	insertAt := func(slice []modfile.Expr, index int, value *modfile.LineBlock) []modfile.Expr {
+		/*
+			Insert inserts the value into the slice at the specified index,
+			which must be in range.
+			from: https://github.com/golang/go/wiki/SliceTricks#insert
+		*/
+		slice = append(slice, nil /* use the zero value of the element type */)
+		copy(slice[index+1:], slice[index:])
+		slice[index] = value
+		return slice
+	}
+
 	if len(lineMods) <= 0 {
 		return
 	}
@@ -99,7 +175,8 @@ func addTestRequireBlock(f *modfile.File, lineMods []lineMod) {
 		Token: []string{"require"},
 		Line:  testLines,
 	}
-	f.Syntax.Stmt = append(f.Syntax.Stmt, newTestBlock)
+	index := findLastRequire(f) + 1
+	f.Syntax.Stmt = insertAt(f.Syntax.Stmt, index, newTestBlock)
 	f.Syntax.Cleanup()
 }